no, really, returning super here doesn't work
I've tested this now a number of times and in a number of ways, it does
not work to return super here in this munge, I'm always getting this error:

```
Error: Failed to apply catalog: Parameter arch failed on Cobblerdistro[ubuntu-12.04.3-x86_64]: Munging failed for value "x86_64" in class arch: implicit argument passing of super from method defined by define_method() is not supported. Specify all arguments explicitly.
```

This patch, fixes the issue.